feat: move `bind` to trait
