+++ src/shims/unix/socket.rs
@@ -5,6 +5,7 @@ use rustc_target::spec::Os;
 use crate::shims::FileDescriptionRef;
 use crate::shims::files::FdNum;
 use crate::shims::unix::UnixFileDescription;
+use crate::shims::unix::socket_address::EvalContextExt as _;
 use crate::shims::unix::tcp_socket::TcpSocket;
 use crate::*;
 
@@ -238,5 +239,31 @@ pub trait EvalContextExt<'tcx>: crate::MiriInterpCxExt<'tcx> {
         let fd = fds.new_ref(TcpSocket::new(family, is_non_block));
 
         interp_ok(Scalar::from_i32(fds.insert(fd)))
+    }
+
+    fn bind(
+        &mut self,
+        socket: &OpTy<'tcx>,
+        address: &OpTy<'tcx>,
+        address_len: &OpTy<'tcx>,
+    ) -> InterpResult<'tcx, Scalar> {
+        let this = self.eval_context_mut();
+
+        let socket = this.read_scalar(socket)?.to_i32()?;
+        let address = match this.read_socket_address(address, address_len, "bind")? {
+            Ok(addr) => addr,
+            Err(e) => return this.set_errno_and_return_neg1_i32(e),
+        };
+
+        // Get the file handle
+        let Some(fd) = this.machine.fds.get(socket) else {
+            return this.set_errno_and_return_neg1_i32(LibcError("EBADF"));
+        };
+
+        let socket = fd.as_unix(this).as_socket(this);
+        match socket.bind(this.machine.communicate(), address, this)? {
+            Ok(_) => interp_ok(Scalar::from_i32(0)),
+            Err(e) => this.set_errno_and_return_neg1_i32(e),
+        }
     }
 }
